+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Improvements:
 Bugfixes:
 
   - Added missing `scopes` parameter in SiteAdapter base class (Josh Johnston)
+  - Deleting authorization token after usage (Josh Johnston)
 
 ## 0.5.0